+++ lib/ogn_client/messages/sender.rb
@@ -14,6 +14,7 @@ module OGNClient
       (?:s(?<flarm_software_version>[\d.]+)\s?)?
       (?:h(?<flarm_hardware_version>[\dA-F]{2})\s?)?
       (?:r(?<flarm_id>[\dA-F]+)\s?)?
+      (?:(?<power>[+-][\d.]+)dBm\s?)?
       (?:hear(?<proximity>.+))?
     $)x
 
@@ -41,14 +42,15 @@ module OGNClient
       3 => :ogn
     }
 
-    attr_reader :stealth_mode             # boolean
-    attr_reader :no_tracking              # boolean
     attr_reader :sender_type              # see SENDER_TYPES
     attr_reader :address_type             # see ADDRESS_TYPES
     attr_reader :id                       # device ID
+    attr_reader :stealth_mode             # boolean
+    attr_reader :no_tracking              # boolean
+    attr_reader :flight_level             # 100 feet QNE
     attr_reader :climb_rate               # meters per second
     attr_reader :turn_rate                # revolutions per minute
-    attr_reader :flight_level             # 100 feet QNE
+    attr_reader :power                    # power ratio in dBm
     attr_reader :signal                   # signal to noise ratio in decibel
     attr_reader :errors                   # number of CRC errors
     attr_reader :frequency_offset         # kilohertz
@@ -63,7 +65,7 @@ module OGNClient
     def parse(raw)
       raw.match SENDER_PATTERN do |match|
         super unless @raw
-        %i(details id climb_rate turn_rate flight_level signal errors frequency_offset gps_accuracy flarm_software_version flarm_hardware_version flarm_id proximity).each do |attr|
+        %i(details id flight_level climb_rate turn_rate power signal errors frequency_offset gps_accuracy flarm_software_version flarm_hardware_version flarm_id proximity).each do |attr|
           send("#{attr}=", match[attr]) if match[attr]
         end
 # NOTE: [@svoop] [ruby21] workaround necessary until support for ruby21 is removed
@@ -85,6 +87,10 @@ module OGNClient
       @id = raw
     end
 
+    def flight_level=(raw)
+      @flight_level = raw.to_f.round(2)
+    end
+
     def climb_rate=(raw)
       @climb_rate = (raw.to_i / 60.0 / 3.2808).round(1)
     end
@@ -93,8 +99,8 @@ module OGNClient
       @turn_rate = (raw.to_i / 4.0).round(1)
     end
 
-    def flight_level=(raw)
-      @flight_level = raw.to_f.round(2)
+    def power=(raw)
+      @power = raw.to_f.round(3)
     end
 
     def signal=(raw)