+++ spec/lib/ogn_client/messages/receiver_spec.rb
@@ -3,25 +3,36 @@ require_relative '../../../spec_helper'
 describe OGNClient::Receiver do
 
   it "must parse valid raw message" do
-    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.4.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm +33.6C RF:+62-0.8ppm/+33.66dB"
+    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.5.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm 4.902V 0.583A +33.6C 14/16Acfts[1h] RF:+62-0.8ppm/+33.66dB/+19.4dB@10km[112619]/+25.0dB@10km[8/15]"
     subject = OGNClient::Message.parse raw
     subject.must_be_instance_of OGNClient::Receiver
     subject.raw.must_equal raw
-    subject.version.must_equal Gem::Version.new('0.2.4')
+    subject.version.must_equal Gem::Version.new('0.2.5')
     subject.platform.must_equal :arm
     subject.cpu_load.must_equal 0.2
     subject.cpu_temperature.must_equal 33.6
+    subject.voltage.must_equal 4.902
+    subject.amperage.must_equal 0.583
+    subject.senders.must_equal 16
+    subject.visible_senders.must_equal 14
+    subject.invisible_senders.must_equal 2
     subject.ram_free.must_equal 777.7
     subject.ram_total.must_equal 972.2
     subject.ntp_offset.must_equal 3.1
     subject.ntp_correction.must_equal -3.8
-    subject.manual_correction.must_equal 62
-    subject.automatic_correction.must_equal -0.8
+    subject.rf_correction_manual.must_equal 62
+    subject.rf_correction_automatic.must_equal -0.8
     subject.signal.must_equal 33.66
+    subject.senders_signal.must_equal 19.4
+    subject.senders_messages.must_equal 112619
+    subject.good_senders_signal.must_equal 25
+    subject.good_and_bad_senders.must_equal 15
+    subject.good_senders.must_equal 8
+    subject.bad_senders.must_equal 7
   end
 
   it "must parse valid raw message without version" do
-    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm +33.6C RF:+62-0.8ppm/+33.66dB"
+    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm 4.902V 0.583A +33.6C 14/16Acfts[1h] RF:+62-0.8ppm/+33.66dB/+19.4dB@10km[112619]/+25.0dB@10km[8/15]"
     subject = OGNClient::Message.parse raw
     subject.must_be_instance_of OGNClient::Receiver
     subject.version.must_be_nil
@@ -29,39 +40,85 @@ describe OGNClient::Receiver do
   end
 
   it "must parse valid raw message without platform" do
-    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.4 CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm +33.6C RF:+62-0.8ppm/+33.66dB"
+    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.5 CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm 4.902V 0.583A +33.6C 14/16Acfts[1h] RF:+62-0.8ppm/+33.66dB/+19.4dB@10km[112619]/+25.0dB@10km[8/15]"
     subject = OGNClient::Message.parse raw
     subject.must_be_instance_of OGNClient::Receiver
     subject.platform.must_be_nil
   end
 
+  it "must parse valid raw message without voltage" do
+    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.5.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm 0.583A +33.6C 14/16Acfts[1h] RF:+62-0.8ppm/+33.66dB/+19.4dB@10km[112619]/+25.0dB@10km[8/15]"
+    subject = OGNClient::Message.parse raw
+    subject.must_be_instance_of OGNClient::Receiver
+    subject.voltage.must_be_nil
+  end
+
+  it "must parse valid raw message without amperage" do
+    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.5.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm 4.902V +33.6C 14/16Acfts[1h] RF:+62-0.8ppm/+33.66dB/+19.4dB@10km[112619]/+25.0dB@10km[8/15]"
+    subject = OGNClient::Message.parse raw
+    subject.must_be_instance_of OGNClient::Receiver
+    subject.amperage.must_be_nil
+  end
+
   it "must parse valid raw message without temperature" do
-    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.4.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm RF:+62-0.8ppm/+33.66dB"
+    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.5.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm 4.902V 0.583A 14/16Acfts[1h] RF:+62-0.8ppm/+33.66dB/+19.4dB@10km[112619]/+25.0dB@10km[8/15]"
     subject = OGNClient::Message.parse raw
     subject.must_be_instance_of OGNClient::Receiver
     subject.cpu_temperature.must_be_nil
   end
 
-  it "must parse valid raw message without signal and corrections" do
-    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.4.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm +33.6C"
+  it "must parse valid raw message without senders" do
+    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.5.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm 4.902V 0.583A +33.6C RF:+62-0.8ppm/+33.66dB/+19.4dB@10km[112619]/+25.0dB@10km[8/15]"
     subject = OGNClient::Message.parse raw
     subject.must_be_instance_of OGNClient::Receiver
-    subject.manual_correction.must_be_nil
-    subject.automatic_correction.must_be_nil
-    subject.signal.must_be_nil
+    subject.visible_senders.must_be_nil
+    subject.invisible_senders.must_be_nil
   end
 
-  it "must parse valid raw message without corrections" do
-    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.4.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm +33.6C RF:+33.66dB"
+  it "must parse valid raw message without signal, corrections nor averages" do
+    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.5.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm 4.902V 0.583A +33.6C 14/16Acfts[1h]"
     subject = OGNClient::Message.parse raw
     subject.must_be_instance_of OGNClient::Receiver
-    subject.manual_correction.must_be_nil
-    subject.automatic_correction.must_be_nil
+    subject.rf_correction_manual.must_be_nil
+    subject.rf_correction_automatic.must_be_nil
+    subject.signal.must_be_nil
+    subject.senders_signal.must_be_nil
+    subject.senders_messages.must_be_nil
+    subject.good_senders_signal.must_be_nil
+    subject.good_senders.must_be_nil
+    subject.bad_senders.must_be_nil
+  end
+
+  it "must parse valid raw message without corrections nor averages" do
+    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.5.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm 4.902V 0.583A +33.6C 14/16Acfts[1h] RF:+33.66dB"
+    subject = OGNClient::Message.parse raw
+    subject.must_be_instance_of OGNClient::Receiver
+    subject.rf_correction_manual.must_be_nil
+    subject.rf_correction_automatic.must_be_nil
     subject.signal.must_equal 33.66
+    subject.senders_signal.must_be_nil
+    subject.senders_messages.must_be_nil
+    subject.good_senders_signal.must_be_nil
+    subject.good_senders.must_be_nil
+    subject.bad_senders.must_be_nil
+  end
+
+  it "must parse valid raw message without averages" do
+    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v0.2.5.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm 4.902V 0.583A +33.6C 14/16Acfts[1h] RF:+62-0.8ppm/+33.66dB"
+    subject = OGNClient::Message.parse raw
+    subject.must_be_instance_of OGNClient::Receiver
+    subject.rf_correction_manual.must_equal 62
+    subject.rf_correction_automatic.must_equal -0.8
+    subject.signal.must_equal 33.66
+    subject.senders_signal.must_be_nil
+    subject.senders_messages.must_be_nil
+    subject.good_senders_signal.must_be_nil
+    subject.good_senders.must_be_nil
+    subject.bad_senders.must_be_nil
   end
 
   it "must raise error if the version is not supported yet" do
-    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v999.999.999.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm +33.6C RF:+62-0.8ppm/+33.66dB"
+    raw = "LKHS>APRS,TCPIP*,qAC,GLIDERN2:/211635h4902.45NI01429.51E&000/000/A=001689 v999.999.999.ARM CPU:0.2 RAM:777.7/972.2MB NTP:3.1ms/-3.8ppm 4.902V 0.583A +33.6C 14/16Acfts[1h] RF:+62-0.8ppm/+33.66dB/+19.4dB@10km[112619]/+25.0dB@10km[8/15]"
     -> { OGNClient::Message.parse(raw) }.must_raise OGNClient::ReceiverError
   end